+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-dev.5] - 2024-08-10
+
+## Changed
+
+- Simplified timer restart.
+
+## Fixed
+
+- Fixed sound attempting to play when set to `none`.
+
+## [1.0.0-dev.4] - 2024-07-31
+
+## Added
+
+- Volume control for timer audio.
+
 ## [1.0.0-dev.3] - 2024-06-23
 
 ## Changed